+++ R/pkg-reimport.R
@@ -11,17 +11,17 @@
 #' pkg_export()
 #' }
 pkg_export <- function(old_lib = "~/Library/R/shared_bak/") {
-  pak::lib_status(lib = old_lib) %>%
-    as_tibble() %>%
-    select(package, title, version, repository, remotetype, remoterepo, remoteusername) %>%
-    glue_data("if (!'{package}' %in% installed.packages()) remotes::install_cran('{package}')") %>%
+  pak::lib_status(lib = old_lib) |>
+    as_tibble() |>
+    select(package, title, version, repository, remotetype, remoterepo, remoteusername) |>
+    glue_data("if (!'{package}' %in% installed.packages()) remotes::install_cran('{package}')") |>
     writeLines("cran-reinstall.R")
 
-  pak::lib_status(lib = old_lib) %>%
-    as_tibble() %>%
-    select(package, title, version, repository, remotetype, remoterepo, remoteusername) %>%
-    filter(remotetype == "github") %>%
-    glue_data("if (!'{package}' %in% installed.packages()) remotes::install_github('{remoteusername}/{remoterepo}')") %>%
+  pak::lib_status(lib = old_lib) |>
+    as_tibble() |>
+    select(package, title, version, repository, remotetype, remoterepo, remoteusername) |>
+    filter(remotetype == "github") |>
+    glue_data("if (!'{package}' %in% installed.packages()) remotes::install_github('{remoteusername}/{remoterepo}')") |>
     writeLines("github-reinstall.R")
 }